refactor(backup): implement consistent timestamp for backup filenames
